requirements.txt
fastapi
uvicorn[standard]
numpy
pydantic
strawberry-graphql[fastapi]
arq
redis
python-multipart
